requirements.txt
absl-py==2.0.0
altair==5.1.2
attrs==23.1.0
blinker==1.6.3
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
Flask==3.0.0
flatbuffers==23.5.26
fonttools==4.43.1
gitdb==4.0.10
GitPython==3.1.37
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
mdurl==0.1.2
mediapipe==0.10.7
numpy==1.26.0
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.1
Pillow==10.0.1
protobuf==3.20.3
pyarrow==13.0.0
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
referencing==0.30.2
requests==2.31.0
rich==13.6.0
rpds-py==0.10.6
six==1.16.0
smmap==5.0.1
sounddevice==0.4.6
streamlit==1.27.2
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
urllib3==2.0.6
validators==0.22.0
Werkzeug==3.0.0
zipp==3.17.0
